+++ aoe4/aoe4_game_overlay.py
@@ -250,6 +250,8 @@ class AoE4GameOverlay(RTSGameOverlay):
         self.config_quit_button.show()
         self.config_save_button.show()
         self.config_reload_button.show()
+        self.config_hotkey_button.show()
+        self.config_build_order_button.show()
         self.font_size_input.show()
         self.scaling_input.show()
         self.next_panel_button.show()
@@ -263,10 +265,6 @@ class AoE4GameOverlay(RTSGameOverlay):
         self.username_title.show()
         self.username_search.show()
         self.username_selection.show()
-
-        # adjust the size of the elements
-        self.build_order_search.adjustSize()
-        self.username_search.adjustSize()
 
         # configuration buttons
         layout = self.settings.layout
@@ -282,6 +280,10 @@ class AoE4GameOverlay(RTSGameOverlay):
         self.config_save_button.move(next_x, border_size)
         next_x += action_button_size + action_button_spacing
         self.config_reload_button.move(next_x, border_size)
+        next_x += action_button_size + action_button_spacing
+        self.config_hotkey_button.move(next_x, border_size)
+        next_x += action_button_size + action_button_spacing
+        self.config_build_order_button.move(next_x, border_size)
         next_x += action_button_size + horizontal_spacing
         self.font_size_input.move(next_x, border_size)
         next_x += self.font_size_input.width() + horizontal_spacing
@@ -804,6 +806,8 @@ class AoE4GameOverlay(RTSGameOverlay):
             self.config_quit_button.hovering_show(self.is_mouse_in_roi_widget)
             self.config_save_button.hovering_show(self.is_mouse_in_roi_widget)
             self.config_reload_button.hovering_show(self.is_mouse_in_roi_widget)
+            self.config_hotkey_button.hovering_show(self.is_mouse_in_roi_widget)
+            self.config_build_order_button.hovering_show(self.is_mouse_in_roi_widget)
 
         elif self.selected_panel == PanelID.BUILD_ORDER:  # build order specific buttons
             self.build_order_previous_button.hovering_show(self.is_mouse_in_roi_widget)